// src/Components/reusableComponents/TableComponent/types.ts
import { MRT_ColumnDef } from "material-react-table";

export interface TableInterface {
    columns?: MRT_ColumnDef[];
    loading?: boolean;
    data?: [] | any;
    handleTable?: (e: any) => any;
    onPageChange?: (pageNumber?: React.ChangeEvent<unknown>, page?: number) => void;
    onPageListChange?: (pageData: any) => void;
    onDownloadCSV?: (csv: any) => void;
    onChangeAction?: (e: any) => void;
    actionDropDown?: string[];
    downloadCSVText?: string;
    downloadCSV?: boolean;
    totalRecord?: number;
    currentPage?: number;
    rowsPerPage?: number;
    pageListOption?: any;
    enablePinning?: boolean;
    enableColumnFilterModes?: boolean;
    enableColumnResizing?: boolean;
    enableFilterMatchHighlighting?: boolean;
    enableGlobalFilterModes?: boolean;
    enableColumnOrdering?: boolean;
    enableStickyHeader?: boolean;
    enableEditing?: boolean;
    enableRowActions?: boolean;
    positionActionsColumn?: "first" | "last";
    editingMode?: "modal" | "cell" | "table" | "row";
    positionToolbarAlertBanner?: "bottom" | "top" | "none";
    icons?: any;
    settings?: any;
    enableColumnActions?: boolean;
    enableColumnFilters?: boolean;
    enablePagination?: boolean;
    enableSorting?: boolean;
    enableBottomToolbar?: boolean;
    enableTopToolbar?: boolean;
    muiTableBodyRowProps?: any;
    muiTableBodyCellProps?: any;
    sorting?: any;
    additionalBtn?: boolean;
    additionalBtnText?: any;
    handleAdditionalFunc?: (e: any) => any;
    toolBarPosition?: "top" | "bottom";
    showFilterButton?: boolean;
    showDownloadButton?: boolean;
    showRefreshButton?: boolean;
    enableMultiRowSelection?: boolean;
    handleFilterButton?: () => void;
    handleRefreshButton?: () => void;
    tableHeader?: any;
}